backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
sqlalchemy==2.0.25
llama-cpp-python==0.2.57
dspy-ai==2.3.0
langsmith==0.1.84
promptlayer==0.2.1
pydantic>=2.7.4,<3.0.0  # This is the magic line that resolves everything
python-multipart==0.0.9
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1
llama-index==0.10.20
transformers==4.40.1
torch==2.2.2
aiosqlite==0.20.0
alembic==1.13.1
pytest-asyncio==0.23.5
httpx==0.27.0